perform_create was not part of PostViewClass - indent error.
The WORST TYPO by far, in posts/views.py you have perform_create() defined at the module level rather than inside of the class PostViewSet.  This is obviously wrong (only took me 8 hours to find it) because it calls super(PostViewSet, self) thus it is supposed to be a member of that class.  With the indent wrong, the javascript returns Error: undefined.